config: Add config file for constants and environment variables
Add config file for centrally monitoring and easier access of environment variables and constants
which will be throughout the code. Variables and constants to be used must be first added to the
config object and then imported in code file needed.
eg: add all environment variables and constants to be used to connect Mongodb under 'config.db' and
    then import in mongoose.js or any other db file.
    config: {
	db: {
	    mongoURI: "mongodb uri needed to connect",
	    //add other key: value pairs if need in mongoose.js or db file
	    ...
	}
	// add other key: value if needed in some other code file
   ...
   }