service: Improve validation for bearer token verification.
Bearer token accepted by the function verifyBearerToken used to directly pass bearertoken string to jwt.verify with even checking if bearertoken contains something and is not null. This also overcomes the situation in a case where user did not provide any token thus value for bearerToken becomes null for this case server must respond HTTP401 with message 'Bearer token not provided' but instead responds with HTTP401 with message 'Not valid Bearer token'.