auth: Improve oauth http error response.
If Passport Google startergy fails to authenticate users due to error in module itself or something went wrong during oauth which failed to generate access token for requested user then user will be notified with HTTP500 error response

Previous behaviour of code will fail to catch these error in ErrorHandler if error are other than mongoDBError or HTTPError which will result in exit of node process due to failure of authention for just some users. Instead now the server will respond with HTTP500 and process will still be running.

Logging of errors must be done if errors are of passport module or if from Google OAuth startergy.